requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP Client
httpx==0.27.2

# HTML Processing
beautifulsoup4==4.12.3
lxml==5.3.0

# Vector Store
faiss-cpu==1.7.4
numpy==1.26.4

# Embeddings (choose one or more)
openai==1.54.3
# google-generativeai==0.8.0  # Uncomment if using Gemini
# sentence-transformers==2.7.0  # Uncomment if using local embeddings

# Utilities
python-multipart==0.0.9

